+++ CHANGELOG.md
@@ -53,7 +53,8 @@ All notable changes to `honeypot` will be documented in this file
 
 
 ## 0.7.0 - 2021-05-04
-- bump sfneal/tracking to v1.0
+- bump sfneal/tracking to v1.0.1
 - start making feature & unit test suites
 - make HoneypotServiceProvider for publishing config files
 - bump phpunit/phpunit & orchestra/testbench composer dev requirements
+- fix issues with missing tracking tokens not being added through middleware